setReferrer and shop

+++ contracts/nft/DayOfRightsClubPackage.sol
@@ -6,23 +6,21 @@ import "../core/SafeOwnable.sol";
 import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
 import "../interfaces/IDayOfRightsClub.sol";
 import "../interfaces/IBlockhashMgr.sol";
+import "hardhat/console.sol";
 
 //ERC721("Day of right club package", "DORCP"),
-contract DayOfRightsClubPackage is
-    ERC721("Test package", "DPT"),
-    SafeOwnable
-{
+contract DayOfRightsClubPackage is ERC721("Test package", "DPT"), SafeOwnable {
     IBlockhashMgr private blockhashMgr;
-    
+
     IDayOfRightsClub private dayOfRightsClub;
     mapping(uint256 => PackageInfo) public packageInfos;
-    mapping(uint => uint) public odds;
+    mapping(uint256 => uint256) public odds;
     mapping(address => bool) public isMinner;
     uint256 public total;
 
     struct PackageInfo {
-        uint blockSeed;
-        uint types;
+        uint256 blockSeed;
+        uint256 types;
     }
 
     event Mint(address account, uint256 tokenId);
@@ -38,7 +36,7 @@ contract DayOfRightsClubPackage is
         odds[4] = 20;
     }
 
-    function setOdds(uint _type, uint _newOdds) external onlyOwner {
+    function setOdds(uint256 _type, uint256 _newOdds) external onlyOwner {
         odds[_type] = _newOdds;
     }
 
@@ -60,7 +58,7 @@ contract DayOfRightsClubPackage is
         emit DelMinner(_minner);
     }
 
-    function mint(address _recipient, uint _type) external onlyMinner {
+    function mint(address _recipient, uint256 _type) external onlyMinner {
         require(
             _recipient != address(0),
             "BridgeCoinClub: recipient is zero address"
@@ -68,10 +66,15 @@ contract DayOfRightsClubPackage is
         require(odds[_type] > 0, "unsupported type");
         total += 1;
         uint256 _tokenId = total;
+
         _mint(_recipient, _tokenId);
+
         packageInfos[_tokenId].blockSeed = block.number + 1;
+
         packageInfos[_tokenId].types = _type;
+
         blockhashMgr.request(packageInfos[_tokenId].blockSeed);
+
         emit Mint(_recipient, _tokenId);
     }
 
@@ -87,7 +90,7 @@ contract DayOfRightsClubPackage is
         _;
     }
 
-    function openPackage(uint tokenId) external {
+    function openPackage(uint256 tokenId) external {
         require(_exists(tokenId), "operator query for nonexistent token");
         require(
             _isApprovedOrOwner(_msgSender(), tokenId),
@@ -101,7 +104,7 @@ contract DayOfRightsClubPackage is
         bytes32 bh = blockhashMgr.getBlockhash(package.blockSeed);
         bytes memory seed = abi.encodePacked(bh, abi.encodePacked(tokenId));
 
-        uint result = (uint256(keccak256(seed)) % (100)) + 1;
+        uint256 result = (uint256(keccak256(seed)) % (100)) + 1;
         if (result <= odds[package.types]) {
             dayOfRightsClub.mint(msg.sender);
         }
